tasks_service/requirements.txt
django>=5.0
djangorestframework
